(PC-17816) refactor(profile): move ContactSupportButton from pages to components
